Module fedimint_atomic_broadcast::db

Structs

SignedBlockKey 🔒

UnitSaver

The UnitSaver enables aleph bft to store its local directed acyclic graph of units on disk in order to recover from a mid session crash. By implementing std::io::Write we allow aleph bft to append bytes to its existing backup similar to a open file in append mode.

UnitsKey 🔒

Functions

complete_session

The function removes the units stored by aleph bft and stores the signed block instead

load_block

open_session

This function loads the aleph bft backup from disk and creates a UnitSaver instance which allows aleph bft to append further bytes to the existing backup