Enum fedimint_atomic_broadcast::Shutdown

pub enum Shutdown {
    Clean(u64),
    MidSession(u64),
}

A clean shutdown can be initiated at the end of every session via the shutdown sender passed to run. This can be used for a coordinated shutdown of a federation in order to upgrade. A mid session shutdown is triggered if the receiver for the OrderedItems is dropped. This mechanism can be used if one wants to shut down a single guardian immediately.

Variants

Clean(u64)

MidSession(u64)