Function fedimint_atomic_broadcast::session::run

pub async fn run(
    index: u64,
    keychain: Keychain,
    backup_loader: Cursor<Vec<u8>>,
    backup_saver: UnitSaver,
    item_receiver: Receiver<Vec<u8>>,
    network_data_receiver: Receiver<Vec<u8>>,
    outgoing_message_sender: Sender<(Message, Recipient)>,
    ordered_item_sender: Sender<(OrderedItem, u64, Sender<Decision>)>,
    signed_block_receiver: Receiver<SignedBlock>
) -> Result<SignedBlock>

this function completes a session with the following steps:

set up the aleph bft session

periodically request the corresponding signed block from peers

combine the accepted items into a block until we reach a preset number of ordered batches or a signed block arrives

collect signatures until we reach the threshold or a signed block arrives